potentially fix ray=1.10 issue
node_visited was returning float features when the observation space expected ints

+++ graphenv/examples/tsp/tsp_nfp_state.py
@@ -72,7 +72,7 @@ class TSPNFPState(TSPState):
         """
         outputs = dict(self.graph_inputs)  # not sure the shallow copy is necessary
 
-        node_visited = np.ones(self.num_nodes)
+        node_visited = np.ones(self.num_nodes, dtype=np.int64)
         node_visited[self.tour] += 1
 
         if len(self.tour) > 1:
